+++ lib/src/main/java/com/diffplug/spotless/json/JacksonJsonConfig.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023-2025 DiffPlug
+ * Copyright 2023-2026 DiffPlug
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -28,6 +28,7 @@ public class JacksonJsonConfig extends JacksonConfig {
 	private static final long serialVersionUID = 1L;
 
 	protected Map<String, Boolean> jsonFeatureToToggle = new LinkedHashMap<>();
+	protected Map<String, Boolean> jsonParserFeatureToToggle = new LinkedHashMap<>();
 
 	// https://github.com/revelc/formatter-maven-plugin/pull/280
 	// By default, Jackson adds a ' ' before separator, which is not standard with most IDE/JSON libraries
@@ -35,6 +36,10 @@ public class JacksonJsonConfig extends JacksonConfig {
 
 	public Map<String, Boolean> getJsonFeatureToToggle() {
 		return Collections.unmodifiableMap(jsonFeatureToToggle);
+	}
+
+	public Map<String, Boolean> getJsonParserFeatureToToggle() {
+		return Collections.unmodifiableMap(jsonParserFeatureToToggle);
 	}
 
 	/**
@@ -49,6 +54,20 @@ public class JacksonJsonConfig extends JacksonConfig {
 	 */
 	public void appendJsonFeatureToToggle(Map<String, Boolean> features) {
 		this.jsonFeatureToToggle.putAll(features);
+	}
+
+	/**
+	 * Refers to com.fasterxml.jackson.core.JsonParser.Feature
+	 */
+	public void setJsonParserFeatureToToggle(Map<String, Boolean> jsonParserFeatureToToggle) {
+		this.jsonParserFeatureToToggle = jsonParserFeatureToToggle;
+	}
+
+	/**
+	 * Refers to com.fasterxml.jackson.core.JsonParser.Feature
+	 */
+	public void appendJsonParserFeatureToToggle(Map<String, Boolean> features) {
+		this.jsonParserFeatureToToggle.putAll(features);
 	}
 
 	public boolean isSpaceBeforeSeparator() {