+++ plugin-maven/src/main/java/com/diffplug/spotless/maven/json/JacksonJson.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023 DiffPlug
+ * Copyright 2023-2026 DiffPlug
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -44,12 +44,16 @@ public class JacksonJson implements FormatterStepFactory {
 	@Parameter
 	private Map<String, Boolean> jsonFeatures = Collections.emptyMap();
 
+	@Parameter
+	private Map<String, Boolean> jsonParserFeatures = Collections.emptyMap();
+
 	@Override
 	public FormatterStep newFormatterStep(FormatterStepConfig stepConfig) {
 		JacksonJsonConfig jacksonConfig = new JacksonJsonConfig();
 
 		jacksonConfig.appendFeatureToToggle(features);
 		jacksonConfig.appendJsonFeatureToToggle(jsonFeatures);
+		jacksonConfig.appendJsonParserFeatureToToggle(jsonParserFeatures);
 		jacksonConfig.setSpaceBeforeSeparator(spaceBeforeSeparator);
 
 		return JacksonJsonStep